Restructure code and module
