+++ back-end/back-end-openai-api/src/modules/open-api/openapi.controller.ts
@@ -1,12 +1,13 @@
 import { Controller, Post, Body } from '@nestjs/common';
 import { OpenApiService } from './openapi.service';
+import { ChatCompletionMessageParam } from 'openai/resources';
 
 @Controller('openapi')
 export class OpenApiController {
   constructor(private readonly openApiService: OpenApiService) {}
 
   @Post('message')
-  async sendMessage(@Body('messages') messages: { role: 'user' | 'assistant' | 'system'; content: string }[]): Promise<any> {
+  async sendMessage(@Body('messages') messages: ChatCompletionMessageParam[]): Promise<any> {
     try {
       const result = await this.openApiService.getOpenApiResponse(messages);
       return { result }; // Trả về phản hồi từ OpenAI
